backend/requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ORM & migrations
sqlalchemy==2.0.29
alembic==1.13.1

# Database driver
psycopg2-binary==2.9.9

# Validation
pydantic==2.6.4
pydantic-settings==2.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Excel I/O
openpyxl==3.1.2
pandas==2.2.1

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
